+++ middleman-core/fixtures/sass-cache-path-custom-app/config.rb
@@ -1,3 +1,0 @@
-
-
-set :sass_cache_path, File.join('/tmp', "#{File.basename(Dir.pwd)}-custom-sass_cache_path")