+++ middleman-core/fixtures/sass-cache-path-default-app/config.rb
@@ -1,3 +1,0 @@
-
-# Using default setting
-# set :sass_cache_path, File.join(Dir.pwd, '.sass-cache')