+++ middleman-core/lib/middleman-core/extension.rb
@@ -1,3 +1,6 @@
+require 'active_support/core_ext/class/attribute'
+require 'active_support/core_ext/module/delegation'
+
 module Middleman
   class Extension
     class_attribute :supports_multiple_instances, :instance_reader => false, :instance_writer => false
@@ -75,6 +78,7 @@ module Middleman
       # Bind app hooks to local methods
       bind_before_configuration
       bind_after_configuration
+      bind_before_build
       bind_after_build
     end
 
@@ -132,6 +136,19 @@ module Middleman
 
         if ext.respond_to?(:manipulate_resource_list)
           ext.app.sitemap.register_resource_list_manipulator(ext.class.extension_name, ext)
+        end
+      end
+    end
+
+    def bind_before_build
+      ext = self
+      if ext.respond_to?(:before_build)
+        @klass.before_build do |builder|
+          if ext.method(:before_build).arity === 1
+            ext.before_build(builder)
+          else
+            ext.before_build
+          end
         end
       end
     end