+++ middleman-core/lib/middleman-core/extensions.rb
@@ -1,6 +1,3 @@
-require 'active_support/core_ext/class/attribute'
-require 'active_support/core_ext/module/delegation'
-
 module Middleman
 
   module Extensions