+++ middleman-core/lib/vendored-middleman-deps/padrino-core-0.12.0/test/fixtures/app_gem/lib/app_gem.rb
@@ -1,0 +1,7 @@
+require 'padrino'
+
+module AppGem
+  extend Padrino::Module
+
+  gem! 'app_gem'
+end