+++ middleman-core/lib/vendored-middleman-deps/padrino-core-0.12.0/test/fixtures/apps/lib/myklass/mysubklass.rb
@@ -1,0 +1,4 @@
+class MyKlass
+    class MySubKlass
+    end
+end